[WIP] Log who will get names
The problem with this is that NAMES is special and will be handled by directly IrcHandler.onJoin-ing each entry in the NAMES list. That is fine, but the users who are NOT on this list should also NOT be on the list provided by the membershiplist syncer. If there are any discrepancies between the remoteJoinedUsers list and the NAMES list, they need to be resolved.

Adding another listener to NAMES will probably not work however...

This is becoming more and more like an architecture problem